removed old service + renamed new one
