+++ src/services/api/activity-log/activity.log.resolver.queries.ts
@@ -5,32 +5,30 @@ import { CurrentUser, Profiling } from '@src/common/decorators';
 import { ActivityLogInput } from './dto/activity.log.dto.collaboration.input';
 import { GraphqlGuard } from '@core/authorization/graphql.guard';
 import { AgentInfo } from '@core/authentication/agent-info';
-import { IActivity } from '@src/platform/activity/activity.interface';
 import { AuthorizationService } from '@core/authorization/authorization.service';
-import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 import { AuthorizationPrivilege } from '@common/enums/authorization.privilege';
 import { CollaborationService } from '@domain/collaboration/collaboration/collaboration.service';
+import { IActivityLogEntry } from './dto/activity.log.entry.interface';
 
 @Resolver()
 export class ActivityLogResolverQueries {
   constructor(
     private activityLogService: ActivityLogService,
     private authorizationService: AuthorizationService,
-    private authorizationPolicyService: AuthorizationPolicyService,
     private collaborationService: CollaborationService
   ) {}
 
   @UseGuards(GraphqlGuard)
-  @Query(() => [IActivity], {
+  @Query(() => [IActivityLogEntry], {
     nullable: false,
     description: 'Retrieve the ActivityLog for the specified Collaboration',
   })
   @Profiling.api
-  async activityLogOnCollaboration(
+  async activityLog2OnCollaboration(
     @CurrentUser() agentInfo: AgentInfo,
     @Args('queryData', { type: () => ActivityLogInput, nullable: false })
     queryData: ActivityLogInput
-  ): Promise<IActivity[]> {
+  ): Promise<IActivityLogEntry[]> {
     const collaboration =
       await this.collaborationService.getCollaborationOrFail(
         queryData.collaborationID
