+++ src/services/api/activity-log/dto/activity.log.dto.entry.base.interface.ts
