+++ src/services/api/activity-log/dto/activity.log.dto.entry.callout.published.ts
