+++ src/services/api/activity-log/dto/activity.log.dto.entry.member.joined.interface.ts
