--- src/services/api/activity-log2/dto/activity.log.dto.collaboration.input.ts
@@ -1,12 +1,0 @@
-import { UUID } from '@domain/common/scalars';
-import { Field, InputType } from '@nestjs/graphql';
-
-@InputType()
-export class ActivityLog2Input {
-  @Field(() => UUID, {
-    nullable: false,
-    description:
-      'Display the activityLog results for the specified Collaboration.',
-  })
-  collaborationID!: string;
-}
