--- src/services/api/activity-log2/index.ts
@@ -1,1 +1,0 @@
-export * from './activity2.log.service';
